+++ src/lib.rs
@@ -1,26 +1,29 @@
 mod assets;
 mod upload;
 
-pub use assets::{AssetIndex, AssetMetadata, KVAssets};
+pub use assets::{init_kv, AssetIndex, AssetMetadata, KVAssets, KV};
 
 // for non-wasm, export asset builders that depend on std::fs and wrangler libs
 #[cfg(not(target_arch = "wasm32"))]
 pub use upload::{sync_assets, SyncConfig};
 
 use thiserror::Error as ThisError;
-#[derive(Debug, ThisError)]
+#[derive(Clone, Debug, ThisError)]
 pub enum Error {
     #[error("KV Api error {0} (body) {1}")]
-    KVHttp(Box<dyn std::error::Error>, String),
+    #[allow(clippy::upper_case_acronyms)]
+    KVHttp(String, String),
 
     #[error("KV Api error Status {0} (body) {1}")]
+    #[allow(clippy::upper_case_acronyms)]
     KVHttpStatus(u16, String),
 
     #[error("KV Key {0} not found. status={1}")]
+    #[allow(clippy::upper_case_acronyms)]
     KVKeyNotFound(String, u16),
 
     #[error("Deserializing assets:{0}")]
-    DeserializeAssets(bincode::Error),
+    DeserializeAssets(String),
 
     #[error("Empty key passed to lookup")]
     EmptyKey,
@@ -31,6 +34,7 @@ pub enum Error {
 
     #[cfg(not(target_arch = "wasm32"))]
     #[error("IO Error: {0}")]
+    #[allow(clippy::upper_case_acronyms)]
     IO(String),
 
     #[cfg(not(target_arch = "wasm32"))]
@@ -50,6 +54,7 @@ pub enum Error {
     MissingWranglerFile(String),
 
     #[error("TTL to short. Must be at least 60 seconds")]
+    #[allow(clippy::upper_case_acronyms)]
     TTLTooShort,
 
     // catch-all